Removed 'fake solvent' code from create_solution, moved the calculations to a new function, _compute_solution_contents(), fixed issues created by 'fake solvent' code
The 'fake solvent' code would only work correctly if the solvent argument did not have any overlap with the solute substances. In the case of a solvent argument that was a container which contained one of the solutes, it would produce an incorrect solution. The internal calculations for the system of equations needed to be re-worked to address this bug. Additionally, to reduce the size of create_solution(), a separate function for the mole calculations, _compute_solution_contents(), was created.