Finished implementing get_mass/moles/volume/quantity() for Contain and adjusted Recipe unit test to temporarily pass
Added an optional substance parameter to get_mass()/get_volume()/get_moles()/get_quantity() and ran them through a custom testing script to ensure correct behavior (this will be made into unit tests in a different branch). Outputs are not rounded at present. \nThe unit test that was changed was previously only passing because of the existence of Recipe.create_container(). By removing that function, it revealed that the intended target of the test, Recipe.remove(), was not behaving properly. This will be fixed in a separate branch at a later time.