Moved Unit.convert() and Unit.convert_from() to Substance
These functions were made into instance methods for Substance, as the first parameter was the substance itself. All other classes/tests which called these methods were updated to use the new call format.

The body of Substance.convert_from() was changed to use a pattern matching structure instead of if-else branching, as it improved the code legibility. This change will need to be reverted for compatibility with Python <=3.9.

Unit is now imported by Substance whereas before it was reversed. To avoid an immediate refactor of the method Unit.convert_from_storage_to_standard_format(), a call to import Substance was added at the top of the method's body.

Due to slight changes to the underlying float arithmetic, two unit tests needed to be changed to include pytest.approx() calls. These tests should be re-examined to determine if a strictly correct answer is desired.