Added Container.dilute_in_place() Unit Test
Added the Unit Test for Container.dilute_in_place() and made a number
of improvements to the Container.create_dilution() unit test. Also made
a number of improvements to the two methods in the Container class
itself to pass the new unit tests. Finally, made some minor adjustments
to Container._transfer()'s error reporting and the docstring/comments in
Container.create_solution(). Example.py was also modified to reflect the
changes to Container.dilute_in_place().