Improved _compute_solution_contents() Unit Test
Adjusted the loops of some success cases in the Unit Test for
Container._compute_solution_contents() to reduce the number of
iterations, thereby reducing the time taken to run the test. The
iterations were reduced by treating the substance permutations and
the unit permutations as independent of each other. The unit
permutations were typically higher in count than the substance
permutations, so the substance permutations were cycled through
until the unit permutations were exhausted.

Ex:
 - Before: 6 substance permutations * 729 unit permutations -> 4374 iterations
 - After: 6 substance permutations and 729 unit permutations -> 729 iterations

This reduced the runtime of the individual unit test from 0.36s to 0.12s
(the runtime with code coverage was reduced from 1.17s to 0.44s).