Added Unit Tests for Container Representations
Added the unit tests for the Container functions __repr__() and
_repr_html_(). These tests only check to ensure that the functions
do not raise errors on the various container fixtures, they do not
check for specific functionality. This commit brings the code coverage
of 'test_Container.py' to 100% (with two exclusions, search for
'pragma: no cover' in Container.py to see them).