Updated Plate/Recipe Remove Methods
Because of the change to the function signature of `Container.remove()`,
the `Plate` and `Recipe` classes' `remove()` methods needed to be
updated to match, as did the method `Recipe.bake()` as well as the call
to `Plate.remove()` in 'Example.py'.

While updating the "remove" operator branch of `Recipe.bake()`, I
realized that there was no type-checking in the method to ensure the
steps of the recipe had both correctly-typed and valid operands. A
new exception class, `RecipeBakeError`, was created to handle cases
where the operation is not valid. This exception is raised when the
operand types or values are not valid for the operation. Currently,
it is only used in the remove section of `Recipe.bake()`, but it will be
expanded to other operations in the near future.