Added Unit Test for Plate.__init__()
As part of implementing the unit test for the Plate constructor, the
type-checking and implementation of the Plate constructor iteslf were
refined. Docstrings and comments were improved, and unexpected edge
cases were handled (e.g. the `bool` types being considered a subtype of
`int`).

The formatting for some of the Container docstrings and Container unit
tests were also cleaned up.