Added More Unit Tests for Plate class
Added the unit tests for `Plate.__getitem__()`, `Plate.__repr__()`,
`Plate.get_volumes()`, `Plate.get_substances()`, and `Plate.get_moles()`
(it should be noted that these are all small unit tests, as most of the
functions are just wrappers for the equivalent function of the
PlateSlicer class).

Also corrected a typo in the docstring of one of the Container unit
tests.