Refactored Remove into Separate Functions
The Container.remove(), Plate.remove(), and PlateSlicer.remove()
functions were refactored into separate functions based on whether a
specific Substance/set of Substances should be removed, or whether a
type of Substance (e.g. all liquids) should be removed. This change
makes the unit tests for the functions cleaner and arguably makes the
API for the end user clearer as well. Unit tests were updated to reflect
these changes.

NOTE: The unit tests for Plate.remove_substances() and
Plate.remove_by_type() are relatively lightweight, as the PlateSlicer
versions of these functions are where the functionality should be tested
robustly. However, it may be worth looking them over and adding extra
"gut check" cases.

In addition to this change, a modification was made to the Container
function Container.create_dilution(). A small tolerance was added to
account for floating-point errors that result in tiny negative volumes
for the solution to the system of equations.